2022 day 11 part 2
this was really hard for me. I tried many different approaches to solving it. One of them was where each item was represented as an array that had the item's value in each of the bases. I never could figure out why it didn't work. Then I looked on reddit for a hint and saw that I should use lcm as the base