tidy up 2022 day 16
part 2 no longer works for the example but for the real input it does. Part 1 + Part 2 ran in 62 seconds for the real input. I think I could make the code tidier since there is a lot of nearly duplicated stuff but I don't want to spend more time on this